backend/requirements.txt
fastapi
uvicorn[standard]
breeze-connect
python-dotenv
pydantic
loguru
psycopg[binary]
SQLAlchemy>=2.0
requests
urllib3
numpy
scipy
pandas
lxml
beautifulsoup4
nselib
nsetools
nsepython
